chore: 데이터 잘 불러오는지 로그 확인
