+++ app/src/main/java/nextstep/github/MainActivity.kt
@@ -1,6 +1,7 @@
 package nextstep.github
 
 import android.os.Bundle
+import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
@@ -10,11 +11,22 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.lifecycle.lifecycleScope
+import kotlinx.coroutines.launch
 import nextstep.github.ui.theme.GithubTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        val appContainer = (application as NextGitHubApplication).appContainer
+        val repository = appContainer.githubRepository
+
+        lifecycleScope.launch {
+            val repos = repository.getRepos()
+            Log.d("TAG", "repos: $repos")
+        }
+
         setContent {
             GithubTheme {
                 // A surface container using the 'background' color from the theme
